initialize - 0x4

+++ srcxx/Platforms/arch-arm64/ARM64Register.cpp
@@ -4,24 +4,23 @@
 
 #include "ARM64Register.h"
 
-
 void DescribeARM64Reigster(ARM64Reg reg, ARM64RegInfo *ri) {
     if (reg >= ARM64_REG_X0 && reg <= ARM64_REG_X28) {
         ri->isInteger = true;
-        ri->width      = 64;
-        ri->meta       = ARM64_REG_X0 + (reg - ARM64_REG_X0);
+        ri->width     = 64;
+        ri->meta      = ARM64_REG_X0 + (reg - ARM64_REG_X0);
     } else if (reg == ARM64_REG_X29 || reg == ARM64_REG_FP) {
         ri->isInteger = true;
-        ri->width      = 64;
-        ri->meta       = ARM64_REG_X29;
+        ri->width     = 64;
+        ri->meta      = ARM64_REG_X29;
     } else if (reg == ARM64_REG_X30 || reg == ARM64_REG_LR) {
         ri->isInteger = true;
-        ri->width      = 64;
-        ri->meta       = ARM64_REG_X30;
+        ri->width     = 64;
+        ri->meta      = ARM64_REG_X30;
     } else if (reg == ARM64_REG_SP) {
         ri->isInteger = true;
-        ri->width      = 64;
-        ri->meta       = ARM64_REG_X31;
+        ri->width     = 64;
+        ri->meta      = ARM64_REG_X31;
     } else {
         ri->index = 0;
         // LOG-NEED
@@ -30,8 +29,8 @@ void DescribeARM64Reigster(ARM64Reg reg, ARM64RegInfo *ri) {
 }
 
 ARM64Reg DisDescribeARM64Reigster(int regIndex, int regWith) {
-    if(regWith == 0)
+    if (regWith == 0)
         regWith = 64;
 
-    return (ARM64Reg)regIndex - ARM64_REG_X0;
+    return (ARM64Reg)(regIndex - (int)ARM64_REG_X0);
 }
